Added safe guards

+++ script/utils/Constants.sol
@@ -202,20 +202,20 @@ contract Constants is Script {
     }
 
     // Circle domain IDs mapping
-    function getCircleDomainId(uint256 chainId) public view returns (uint256) {
-        int256 cctpDomain = _getCctpDomain(chainId);
+    function getCircleDomainId(uint256 chainId) public view returns (uint32) {
+        int32 cctpDomain = _getCctpDomain(chainId);
         if (cctpDomain == -1) {
             revert("Circle domain ID not found");
         }
-        return uint256(cctpDomain);
+        return uint32(cctpDomain);
     }
 
     function hasCctpDomain(uint256 chainId) public view returns (bool) {
         return _getCctpDomain(chainId) != -1;
     }
 
-    function _getCctpDomain(uint256 chainId) internal view returns (int256) {
-        return vm.parseJsonInt(file, string.concat(".PUBLIC_NETWORKS.", vm.toString(chainId), ".cctpDomain"));
+    function _getCctpDomain(uint256 chainId) internal view returns (int32) {
+        return int32(vm.parseJsonInt(file, string.concat(".PUBLIC_NETWORKS.", vm.toString(chainId), ".cctpDomain")));
     }
 
     function getOftEid(uint256 chainId) public view returns (uint256) {
